cleared formatting bugs

+++ wk3/collinear/Point.java
@@ -66,7 +66,7 @@ public class Point implements Comparable<Point> {
             return Double.POSITIVE_INFINITY;
         if (this.y - that.y == 0)
             return 0.0;
-        return (that.y - this.y) / (that.x - this.x);
+        return (double) (that.y - this.y) / (that.x - this.x);
     }
 
     /**
